requirements.txt
discord
python-dotenv
asyncio
requests
beautifulsoup4